created daemon/channel
* created daemon/channel
* moved the channel connection logic from daemon/state to daemon/channel
