+++ boot-core/Cargo.toml
@@ -15,8 +15,10 @@ exclude = [".env"]
 [features]
 default = ["daemon"]
 # enable the optional dependencies
+
 daemon = [
     "dep:ibc-chain-registry",
+    "dep:ibc-relayer-types",
     "dep:tokio",
     "dep:tonic",
     "dep:secp256k1",
@@ -58,6 +60,7 @@ bitcoin                 = { version = "0.30.0", optional = true }
 hex                     = { version = "0.4.3", optional = true }
 ripemd                  = { version = "0.1.3", optional = true }
 ibc-chain-registry      = { version = "0.23.0", optional = true }
+ibc-relayer-types       = { version = "0.23.0", optional = true }
 tokio                   = { version = "1.4", features = ["full"], optional = true }
 tonic                   = { version = "0.8.1", optional = true, features = ["tls", "tls-roots"] }
 secp256k1               = { version = "0.27.0", default-features = false, optional = true }
