+++ boot-core/src/daemon/mod.rs
@@ -1,3 +1,4 @@
+pub mod channel;
 pub mod contract;
 pub mod core;
 pub mod error;
