+++ boot-core/src/daemon/state.rs
@@ -1,10 +1,10 @@
 use super::error::DaemonError;
+use crate::daemon::channel::DaemonChannel;
 use crate::error::BootError;
 use crate::state::StateInterface;
-use cosmrs::{
-    proto::cosmos::base::tendermint::v1beta1::{service_client::ServiceClient, GetNodeInfoRequest},
-    Denom,
-};
+
+use cosmrs::Denom;
+
 use cosmwasm_std::Addr;
 use ibc_chain_registry::chain::{Apis, ChainData as RegistryChainInfo, FeeToken, FeeTokens, Grpc};
 use serde::{Deserialize, Serialize};
@@ -17,7 +17,7 @@ use std::{
     rc::Rc,
     str::FromStr,
 };
-use tonic::transport::{Channel, ClientTlsConfig};
+use tonic::transport::Channel;
 
 pub const DEFAULT_DEPLOYMENT: &str = "default";
 
@@ -70,7 +70,9 @@ impl DaemonState {
     pub async fn new(options: DaemonOptions) -> Result<DaemonState, DaemonError> {
         let network: RegistryChainInfo = options.network;
 
-        let mut successful_connections = vec![];
+        let successful_connections = DaemonChannel::new(&network.apis.grpc, &network.chain_id)
+            .await?
+            .unwrap();
 
         if network.apis.grpc.is_empty() {
             return Err(DaemonError::GRPCListIsEmpty);
@@ -79,69 +81,6 @@ impl DaemonState {
         log::info!("Found {} gRPC endpoints", network.apis.grpc.len());
 
         // find working grpc channel
-        for Grpc { address, .. } in network.apis.grpc.iter() {
-            log::info!("Trying to connect to endpoint: {}", address);
-
-            // get grpc endpoint
-            let endpoint = Channel::builder(address.clone().try_into().unwrap());
-
-            // try to connect to grpc endpoint
-            let maybe_client = ServiceClient::connect(endpoint.clone()).await;
-
-            // connection succeeded
-            let mut client = if maybe_client.is_ok() {
-                maybe_client?
-            } else {
-                log::warn!(
-                    "Cannot connect to gRPC endpoint: {}, {:?}",
-                    address,
-                    maybe_client.unwrap_err()
-                );
-
-                // try HTTPS approach
-                // https://github.com/hyperium/tonic/issues/363#issuecomment-638545965
-                if !(address.contains("https") || address.contains("443")) {
-                    continue;
-                };
-
-                log::info!("Attempting to connect with TLS");
-
-                // re attempt to connect
-                let endpoint = endpoint.clone().tls_config(ClientTlsConfig::new())?;
-                let maybe_client = ServiceClient::connect(endpoint.clone()).await;
-
-                // connection still fails
-                if maybe_client.is_err() {
-                    log::warn!(
-                        "Cannot connect to gRPC endpoint: {}, {:?}",
-                        address,
-                        maybe_client.unwrap_err()
-                    );
-                    continue;
-                };
-
-                maybe_client?
-            };
-
-            // get client information for verification down below
-            let node_info = client
-                .get_node_info(GetNodeInfoRequest {})
-                .await?
-                .into_inner();
-
-            // verify we are connected to the spected network
-            if node_info.default_node_info.as_ref().unwrap().network != network.chain_id.as_str() {
-                log::error!(
-                    "Network mismatch: connection:{} != config:{}",
-                    node_info.default_node_info.as_ref().unwrap().network,
-                    network.chain_id.as_str()
-                );
-                continue;
-            }
-
-            // add endpoint to succesful connections
-            successful_connections.push(endpoint.connect().await?)
-        }
 
         // we could not get any succesful connections
         if successful_connections.is_empty() {
